requirements.txt
# === Core Dependencies ===
fastapi>=0.100.0,<0.101.0
uvicorn>=0.20.0,<0.22.0
python-dotenv>=1.0.0,<1.1.0
pydantic>=2.0.0,<2.2.0

# === LangChain & AI ===
langchain>=0.0.300,<0.0.350
langchain-core>=0.1.0,<0.2.0
langchain-community>=0.2.17,<0.3.0
langchain-openai>=0.1.25,<0.2.0
langchain-huggingface>=0.0.3,<0.1.0
openai>=1.52.0,<1.60.0
tiktoken>=0.7.0,<0.8.0

# === Vector Database & Search ===
chromadb>=0.4.0,<0.5.0
sqlalchemy>=2.0.0,<2.2.0
rank-bm25>=0.2.2,<0.3.0

# === Data Processing ===
numpy>=1.24.0,<1.27.0
pandas>=2.0.0,<2.2.0
beautifulsoup4>=4.12.3,<4.14.0
PyMuPDF>=1.24.14,<1.26.0

# === UI / Frontend ===
streamlit>=1.39.0,<1.41.0

# === Utils ===
tqdm>=4.65.0,<4.70.0

# === Re-ranker ===
FlagEmbedding>=1.3.2,<1.4.0

# === Groq ===
groq